Fix ledger query-mode
Query mode is for running a ledger without a transactor, just a query server. The
problem was that we were initializing the query server connection with a custom
:storage-read function that would look to the local ledger server for data, instead of
the remote ledger group. This was fixed by relying on the fluree db's default
:storage-read function, which properly connects to the remote ledger group.

In query mode we were also throwing periodic exceptions as we were trying to collect
ledger stats from the (nonexistant) local ledger group. Now we initiate-stats-reporting
only when we are starting a transactor.

I've also added a couple of additional checks so that users do not need to specify
:fdb-group-servers in the config for a query-mode peer, since that is unused in this
mode.