Enable the local group
This removes the conditionals that check if a peer is a transactor and omit group
initialization. A query peer needs to be able to transfer blocks from a source "group" -
usually a ledger group, but could be another query group.

This commit doesn't actually add any new functionality, it restores us back to how
things worked before I tried to "fix" them. The important thing that was missing before
was group configuration:
- fdb-group-servers
- fdb-group-this-server