Fix shared state between servers in the same jvm
Running two (or n) servers in the same repl is quite useful for development. However, we
had some shared state that was getting clobbered - we were losing the handle to the tcp
server and the http server so that we couldn't gracefully shut them down.

This fixes that problem by changing the atoms from singletons `(atom <singleton-state>)` to maps (atom
{<server-specific-key> <singleton-state}).

We're using the port as a key for the http server(s) and the server-id as the key for
the tcp server's event-loop, just because those keys are handy where we need them.