Fix block key construction so we can find blocks correctly
When we renamed "dbid" to "ledger", we ended up with two meanings for ledger:
1. the tail part of a ledger id
2. the whole ledger id

We need to distinguish between these cases because of how we compose strings. So I've
reverted to using "dbid" to identify the second part of a ledger id. This is only
visible within the codebase and no user will ever see "dbid" in an error message or
otherwise.

Fixes #FC-1414