Ensure retracted subject flakes start with the retracted subject
When we retract a subject we also retract all references to the subject. This can cause
confusion downstream as we often check the first flake's subject to find the subject id,
and we'd see the subject of some other flake that references the subject we're
deleting.

By concatting refs onto flakes instead of the other way around we can ensure that
the `(first subject-flakes)` idiom will return the subject being retracted.